remove cast

+++ two-bulb-imp/main.cpp
@@ -45,7 +45,7 @@ int main(int argc, const char * argv[]) {
     t_params.to = 0.0; // Initial time (h)
     t_params.tf = 20.0; // Final time (h)
     t_params.nt = 5; // Number of time steps.
-    t_params.dt = (double) (t_params.tf - t_params.to) / t_params.nt; // Time sampling
+    t_params.dt = (t_params.tf - t_params.to) / t_params.nt; // Time sampling
     
     // Diffusivities
     p_params.D12 = 8.33e-5 * 3600; // units (m2 / h)
